+++ math/benches/utils/metal_functions.rs
@@ -1,6 +1,6 @@
 use lambdaworks_gpu::metal::abstractions::state::MetalState;
+use lambdaworks_math::fft::gpu::metal::ops::*;
 use lambdaworks_math::fft::polynomial::FFTPoly;
-use lambdaworks_math::gpu::metal::fft::ops::*;
 use lambdaworks_math::{field::traits::RootsConfig, polynomial::Polynomial};
 
 // WARN: These should always be fields supported by Metal, else the last two benches will use CPU FFT.
