+++ math/benches/utils/stark252_utils.rs
@@ -1,4 +1,5 @@
 use lambdaworks_math::{
+    fft::cpu::roots_of_unity::get_twiddles,
     field::{
         element::FieldElement, fields::fft_friendly::stark_252_prime_field::Stark252PrimeField,
         traits::RootsConfig,
@@ -28,5 +29,5 @@ pub fn rand_poly(order: u64) -> Polynomial<FE> {
 #[no_mangle]
 #[export_name = "util::get_twiddles"]
 pub fn twiddles(order: u64, config: RootsConfig) -> Vec<FE> {
-    lambdaworks_math::fft::roots_of_unity::get_twiddles(order, config).unwrap()
+    get_twiddles(order, config).unwrap()
 }
