+++ math/src/fft/cpu/fft.rs
@@ -98,10 +98,9 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::fft::{
-        test_helpers::naive_matrix_dft_test,
-        {bit_reversing::in_place_bit_reverse_permute, roots_of_unity::get_twiddles},
-    };
+    use crate::fft::cpu::bit_reversing::in_place_bit_reverse_permute;
+    use crate::fft::cpu::roots_of_unity::get_twiddles;
+    use crate::fft::test_helpers::naive_matrix_dft_test;
     use crate::field::{test_fields::u64_test_field::U64TestField, traits::RootsConfig};
     use proptest::{collection, prelude::*};
