+++ math/src/fft/gpu/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "metal")]
+pub mod metal;
+
+#[cfg(feature = "cuda")]
+pub mod cuda;
