+++ math/src/fft/mod.rs
@@ -1,9 +1,7 @@
-pub mod bit_reversing;
+pub mod cpu;
 pub mod errors;
-pub mod fft_iterative;
-pub mod ops;
+pub mod gpu;
 pub mod polynomial;
-pub mod roots_of_unity;
 
 #[cfg(test)]
 pub(crate) mod test_helpers;
