--- math/src/gpu/metal/mod.rs
@@ -1,1 +1,0 @@
-pub mod fft;
